+++ JobApplication/src/main/java/cr/ac/una/job/exceptions/OferenteNotFoundException.java
@@ -1,7 +1,7 @@
 package cr.ac.una.job.exceptions;
 
 public class OferenteNotFoundException extends RuntimeException {
-    public OferenteNotFoundException(Integer id) {
+    public OferenteNotFoundException(Long id) {
         super("Oferente con id " + id + " no encontrado");
     }
 }
