+++ JobApplication/src/main/java/cr/ac/una/job/exceptions/PuestoNotFoundException.java
@@ -1,7 +1,7 @@
 package cr.ac.una.job.exceptions;
 
 public class PuestoNotFoundException extends RuntimeException {
-    public PuestoNotFoundException(Integer id) {
+    public PuestoNotFoundException(Long id) {
         super("Puesto con id " + id + " no encontrado");
     }
 }
